requirements.txt
aiohttp==3.11.18
chromadb==1.0.8
fastapi==0.115.9
langchain_mcp_adapters==0.1.11
langchain_openai==1.0.1
langgraph==1.0.2
markdownify==1.1.0
mcp==1.19.0
openai==2.6.1
pydantic==2.11.4
python-dotenv==1.1.0
readabilipy==0.3.0
uvicorn==0.34.2
plotly==6.1.1
filetype==1.2.0
gradio==5.49.1
posthog<6.0.0
google-genai==1.47.0
langchain==1.0.3
langchain-google-genai==3.0.1